Refactor movement a bit.
This made some methods to help reduce the overall size of input system.
It also moved the run vs walk component out of input system and pushed
it into collision.  This will allow future mobile things to adjust for
their run velocity based on a component which specifies it (not there
yet).